dept(typescript) : migrate info command to typescript
