+++ commands/command.loader.ts
@@ -7,6 +7,6 @@ export class CommandLoader {
   public static load(program): void {
     new NewCommand(new NewAction()).load(program);
     new GenerateCommand(new GenerateAction()).load(program);
-    // new InfoCommand(new InfoAction()).load(program);
+    new InfoCommand(new InfoAction()).load(program);
   }
 }
