+++ commands/info.command.ts
@@ -2,6 +2,9 @@ import { AbstractCommand } from './abstract.command';
 
 export class InfoCommand extends AbstractCommand {
   public load(program: any) {
-    throw new Error("Method not implemented.");
+    program
+      .command('info', 'Display Nest CLI details')
+      .alias('i')
+      .action(this.action.handle);
   }
 }
