--- commands/info.js
@@ -1,6 +1,0 @@
-module.exports = (program) => {
-  program
-    .command('info', 'Display Nest CLI details')
-    .alias('i')
-    .action(require('../actions/info'));
-};
