+++ lib/package-managers/package-manager.factory.ts
@@ -1,10 +1,11 @@
-import { readdir } from "fs";
-import { PackageManager } from "./package-manager";
-import { NpmPackageManager } from "./npm.package-manager";
-import { YarnPackageManager } from "./yarn.package-manager";
+import { readdir } from 'fs';
+import { PackageManager } from './package-manager';
+import { AbstractPackageManager } from './abstract.package-manager';
+import { NpmPackageManager } from './npm.package-manager';
+import { YarnPackageManager } from './yarn.package-manager';
 
 export class PackageManagerFactory {
-  public static create(name, logger) {
+  public static create(name, logger): AbstractPackageManager {
     switch (name) {
       case PackageManager.NPM:
         return new NpmPackageManager(logger);
@@ -13,8 +14,8 @@ export class PackageManagerFactory {
     }
   }
 
-  public static find(logger) {
-    return new Promise((resolve) => {
+  public static async find(logger): Promise<AbstractPackageManager> {
+    return new Promise<AbstractPackageManager>((resolve) => {
       readdir(process.cwd(), (error, files) => {
         if (error) {
           resolve(this.create(PackageManager.NPM, logger));
