+++ package.json
@@ -38,6 +38,7 @@
     "@types/node": "^9.6.6",
     "@types/node-emoji": "^1.8.0",
     "@types/ora": "^1.3.4",
+    "@types/os-name": "^2.0.0",
     "typescript": "^2.8.3"
   }
 }
